requirements.txt
absl-py==0.8.1
aiml==0.9.2
apturl==0.5.2
asgiref==3.2.7
asn1crypto==0.24.0
astor==0.8.1
astroid==2.1.0
attrs==19.3.0
audioread==2.1.8
autocorrect==0.4.4
backcall==0.1.0
bcrypt==3.1.6
biopython==1.76
biotite==0.19.2
bleach==3.1.0
blinker==1.4
boto==2.49.0
boto3==1.10.44
botocore==1.13.44
Brlapi==0.6.7
certifi==2018.8.24
cffi==1.14.0
chardet==3.0.4
click==7.1.1
command-not-found==0.3
cryptography==2.6.1
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.12
decorator==4.4.1
defer==1.0.6
defusedxml==0.6.0
distro==1.3.0
distro-info===0.21ubuntu4
Django==3.0.4
django-crispy-forms==1.9.0
dnspython==1.16.0
docutils==0.15.2
duplicity==0.8.4
easydict==1.9
entrypoints==0.3
et-xmlfile==1.0.1
fastdtw==0.3.4
fasteners==0.12.0
Flask==1.1.1
future==0.16.0
gast==0.3.2
gensim==3.8.1
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
httplib2==0.11.3
idna==2.6
imbalanced-learn==0.6.1
imblearn==0.0
importlib-metadata==1.5.0
indicator-cpufreq==0.2.2
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
isort==4.3.4
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.9.4
joblib==0.14.1
json5==0.9.1
jsonschema==3.2.0
jupyter-client==5.3.4
jupyter-core==4.6.2
jupyterlab==1.2.6
jupyterlab-server==1.0.6
kazam==1.4.5
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==18.0.1
keyrings.alt==3.1.1
kiwisolver==1.1.0
language-selector==0.1
launchpadlib==1.10.7
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.3.1
libconfig==0.9.2
librosa==0.7.2
llvmlite==0.31.0
lockfile==0.12.2
logilab-common==1.4.3
louis==3.10.0
luminol==0.4
macaroonbakery==1.2.3
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
mistune==0.8.4
monotonic==1.5
msgpack==0.6.2
mysql-connector-python==8.0.19
nbconvert==5.6.1
nbformat==5.0.4
netifaces==0.10.4
networkx==2.4
nltk==3.4.5
noisereduce==1.1.0
notebook==6.0.3
numba==0.48.0
numpy==1.17.4
oauthlib==2.1.0
olefile==0.46
openpyxl==3.0.3
pandas==0.25.3
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.6.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==6.1.0
ProDy==1.10.11
prometheus-client==0.7.1
prompt-toolkit==3.0.3
PROPKA==3.1.0
protobuf==3.6.1
ptyprocess==0.6.0
PyAudio==0.2.11
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
pydub==0.23.1
pyflakes==2.1.1
pygame==1.9.6
Pygments==2.5.2
PyGObject==3.34.0
PyJWT==1.7.0
pylint==2.2.2
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.5
pyRFC3339==1.1
pyrsistent==0.15.7
python-aiml==0.9.3
python-apt==1.9.0+ubuntu1.3
python-daemon==2.2.4
python-dateutil==2.7.3
python-debian==0.1.36
python-vlc==3.0.7110
pytz==2019.2
pyxdg==0.25
PyYAML==5.1.2
pyzmq==18.1.1
reportlab==3.5.23
requests==2.21.0
requests-unixsocket==0.1.5
resampy==0.2.2
rstoolbox==1.0.1
s3transfer==0.2.1
schrodinger==0.1
scikit-learn==0.22
scipy==1.3.3
seaborn==0.9.0
SecretStorage==2.3.1
selenium==3.141.0
Send2Trash==1.5.0
simplejson==3.16.0
six==1.12.0
sklearn==0.0
smart-open==1.9.0
SoundFile==0.10.3.post1
soundmeter==0.1.5
sqlparse==0.3.1
system-service==0.3
systemd-python==234
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
torch==1.3.1
torchvision==0.4.2
tornado==6.0.3
tqdm==4.42.1
traitlets==4.3.2
ubuntu-advantage-tools==19.5
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.24.1
usb-creator==0.3.7
virtualenv==15.1.0
wadllib==1.3.3
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==0.16.0
wget==3.2
wrapt==1.11.2
xkit==0.0.0
xlrd==1.2.0
zipp==2.2.0
